pyproject.toml
[tool.poetry]
name = "ml-lightning-template"
version = "0.1.0"
description = ""
authors = ["martinHwang <dhhwang89@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.10.2"
omegaconf = "^2.2.3"
PyYAML = "^6.0"
dependency-injector = "^4.40.0"
torch = {url = "https://download.pytorch.org/whl/cpu/torch-1.12.1-cp38-none-macosx_10_9_x86_64.whl", markers = "sys_platform == 'darwin'"}
torchvision = {url = "https://download.pytorch.org/whl/torchvision-0.13.1-cp38-cp38-macosx_10_9_x86_64.whl", markers = "sys_platform == 'darwin'"}
torchaudio = "0.12.1"
pytorch-lightning = "^1.8.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
